requirements.txt
asgiref==3.5.2
Django==3.2
django-jazzmin==2.6.0
django-localflavor==3.1
python-stdnum==1.18
pytz==2022.6
sqlparse==0.4.3
django-bootstrap-v5
django_bootstrap_icons
django-middleware-global-request==0.3.1
# git+https://github.com/whanderley/templated-docs.git@master
# django-fixture-magic
whitenoise
psycopg2
ipython
tqdm
django-ajax-selects
python-dateutil
# Para criar admins de configurações que só pode ter um objeto por classe
#django-solo==2.0.0
#